sm8150-common: biometrics/fod: Uprev to 2.3
 * [SebaUbuntu]: Move fingerprint HIDL to 2.3
 * Remove Lineage FOD HIDL
 * Rename fodextension to udfpsextension
