+++ biometrics/android.hardware.biometrics.fingerprint@2.3-service.xiaomi_msmnile.rc
@@ -1,4 +1,4 @@
-service vendor.fps_hal /vendor/bin/hw/android.hardware.biometrics.fingerprint@2.1-service.xiaomi_msmnile
+service vendor.fps_hal /vendor/bin/hw/android.hardware.biometrics.fingerprint@2.3-service.xiaomi_msmnile
     # "class hal" causes a race condition on some devices due to files created
     # in /data. As a workaround, postpone startup until later in boot once
     # /data is mounted.
@@ -42,6 +42,9 @@ on boot
     chown system system /sys/class/drm/card0-DSI-1/panel_info
     chmod 0744 /sys/class/drm/card0-DSI-1/panel_info
     chmod 0666 /dev/input/event2
+
+    chown system system /sys/devices/platform/soc/soc:qcom,dsi-display-primary/fod_ui
+    chmod 0444 /sys/devices/platform/soc/soc:qcom,dsi-display-primary/fod_ui
 
 on post-fs-data
     mkdir /data/vendor/fpc 0770 system system
