--- fod/vendor.lineage.biometrics.fingerprint.inscreen@1.0-service.xiaomi_msmnile.xml
@@ -1,11 +1,0 @@
-<manifest version="1.0" type="device">
-    <hal format="hidl">
-        <name>vendor.lineage.biometrics.fingerprint.inscreen</name>
-        <transport>hwbinder</transport>
-        <version>1.0</version>
-        <interface>
-            <name>IFingerprintInscreen</name>
-            <instance>default</instance>
-        </interface>
-    </hal>
-</manifest>
