+++ msmnile.mk
@@ -198,13 +198,8 @@ PRODUCT_PACKAGES += \
 endif
 
 # Fingerprint
-ifeq ($(TARGET_HAS_FOD),true)
 PRODUCT_PACKAGES += \
-    vendor.lineage.biometrics.fingerprint.inscreen@1.0-service.xiaomi_msmnile
-endif
-
-PRODUCT_PACKAGES += \
-    android.hardware.biometrics.fingerprint@2.1-service.xiaomi_msmnile
+    android.hardware.biometrics.fingerprint@2.3-service.xiaomi_msmnile
 
 PRODUCT_PACKAGES += \
     vendor.xiaomi.hardware.fingerprintextension@1.0.vendor
